% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{item}
\alias{item}
\title{create items class}
\usage{
item(name, sell_in, quality)
}
\arguments{
\item{name}{A name for the item}

\item{sell_in}{A numeric denoting sell by in days}

\item{quality}{quality}
}
\value{
An items object
}
\description{
create items class
}
\examples{
new_item <- item("laptop", 0, 0)
}
